fix: deslop - reorganize components and fix lint issues
- Reorganize file-tree component into modular folder structure
- Reorganize course-info component into modular folder structure
- Remove AI-generated comments and simplify logic
- Add knip for dead code detection
- Fix knip lint issues
