+++ .prettierignore
@@ -1,1 +1,5 @@
 components/ui
+content/
+pnpm-lock.yaml
+.github/
+scripts/
