+++ components/course-info/index.ts
@@ -1,0 +1,1 @@
+export { CourseInfo } from './course-info';
