+++ components/course-info/info-item.tsx
@@ -1,0 +1,21 @@
+import type { ReactNode, ComponentType, SVGProps } from 'react';
+
+export function InfoItem({
+  label,
+  value,
+  icon: Icon,
+}: {
+  label: string;
+  value: ReactNode;
+  icon: ComponentType<SVGProps<SVGSVGElement>>;
+}) {
+  return (
+    <div className="flex items-start gap-3">
+      <dt className="text-muted-foreground flex items-center gap-2 text-sm">
+        <Icon className="size-4" />
+        {label}
+      </dt>
+      <dd className="text-sm font-medium">{value}</dd>
+    </div>
+  );
+}
