+++ components/course-info/utils.ts
@@ -1,0 +1,3 @@
+export function formatCredit(credit: number) {
+  return Number.isInteger(credit) ? credit.toFixed(1) : credit.toString();
+}
