+++ components/file-tree/index.ts
@@ -1,3 +1,1 @@
-export { Files, Folder, File, FileTreeFromData } from './files';
-export { FileTreeTable as DataTable } from './table';
-export type { FileNode } from '@/lib/types';
+export { Files, Folder, File } from './files';
