+++ lib/source.tsx
@@ -18,14 +18,6 @@ export function getPageImage(page: InferPageType<typeof source>) {
   };
 }
 
-export async function getLLMText(page: InferPageType<typeof source>) {
-  const processed = await page.data.getText('processed');
-
-  return `# ${page.data.title}
-
-${processed}`;
-}
-
 export const blog = loader({
   baseUrl: '/blog',
   source: blogPosts.toFumadocsSource(),
