+++ mdx-components.tsx
@@ -3,7 +3,7 @@ import defaultMdxComponents from 'fumadocs-ui/mdx';
 import type { MDXComponents } from 'mdx/types';
 import { CourseInfo } from '@/components/course-info';
 import type { CourseInfoData } from '@/lib/types';
-import { Files, Folder, File, FileTreeFromData } from '@/components/file-tree';
+import { Files, Folder, File } from '@/components/file-tree';
 import { Accordion, Accordions } from 'fumadocs-ui/components/accordion';
 
 type MdxContext = {
@@ -19,7 +19,6 @@ export function getMDXComponents(
     Files,
     Folder,
     File,
-    FileTreeFromData,
     Accordion,
     Accordions,
     CourseInfo: (props: ComponentProps<typeof CourseInfo>) => (
