+++ package.json
@@ -10,11 +10,11 @@
     "postinstall": "fumadocs-mdx",
     "lint": "eslint .",
     "format": "prettier --write .",
-    "prepare": "husky"
+    "prepare": "husky",
+    "knip": "knip"
   },
   "dependencies": {
     "@radix-ui/react-checkbox": "^1.3.3",
-    "@radix-ui/react-label": "^2.1.8",
     "@radix-ui/react-slot": "^1.2.4",
     "@radix-ui/react-switch": "^1.2.6",
     "@tanstack/react-table": "^8.21.3",
@@ -36,7 +36,6 @@
   "devDependencies": {
     "@commitlint/cli": "^20.3.0",
     "@commitlint/config-conventional": "^20.3.0",
-    "@eslint/eslintrc": "^3.3.3",
     "@tailwindcss/postcss": "^4.1.18",
     "@types/mdx": "^2.0.13",
     "@types/node": "^24.10.4",
@@ -46,11 +45,11 @@
     "eslint-config-next": "16.0.8",
     "eslint-config-prettier": "^10.1.8",
     "husky": "^9.1.7",
+    "knip": "^5.82.1",
     "lint-staged": "^16.2.7",
     "postcss": "^8.5.6",
     "prettier": "3.7.4",
     "prettier-plugin-tailwindcss": "^0.7.2",
-    "shadcn": "^3.6.3",
     "tailwindcss": "^4.1.18",
     "tw-animate-css": "^1.4.0",
     "typescript": "^5.9.3"
